Csv downloads (#3)
* Adding CSV download option

The direct download link is currently broken. I don't want to just
read the CSV from the disc again, as that would break if I move to a
different backing store. Instead, I'll convert what I get from the
store to CSV.

* Defaults to responding with CSV data

Since direct links will be without a decent "Accepts" header, we
default to CSV.

* Altered suffix handler

I'd prefer to only invoke the handler if the URL matches one of the
suffixes, but that's still going to get executed for every request.

* Format response handling now up to snuff

There's too much repetition, but I'll extend the response object
at a later date.